Refactoring component after introducing revert changes button.
