+++ test/unit/index.js
@@ -46,6 +46,7 @@ require('./modules/persistence/rest_spec.js');
 require('./components/explorer/index_spec.js');
 require('./components/explorer/csv_extraction_spec.js');
 require('./components/explorer/query_builder/index_spec.js');
+require('./components/explorer/query_builder/builder_buttons_spec.js');
 require('./components/explorer/visualization/index_spec.js');
 require('./components/explorer/visualization/chart_spec.js');
 require('./components/explorer/saved_queries/browse_queries_spec.js');
@@ -61,7 +62,6 @@ require('./components/common/timezone_spec.js');
 require('./components/common/datepicker_spec.js');
 require('./components/common/relative_picker_spec.js');
 require('./components/common/absolute_picker_spec.js');
-require('./components/common/run_query_spec.js');
 require('./components/common/modal_spec.js');
 require('./components/common/fields_toggle_spec.js');
 require('./components/common/event_browser_spec.js');
